Improve shuffle logic (#83)
* Respect legendaries to legandaries flag if not shadow

* Boost XD Shadow levels
Fixes #79

* Add test base
Add trainer tests

* Fix bugged move tutor definition
Fixes #77

* Refactor pokemon to share same base interface
Used to funnel pokemon randomization through the same set of methods
Add "NextElement" to RNG methods

* Improve string handling and string tables
Update TM description with new move dewscription

* Add Areas and Rooms to extractor

* Fix both Duking and Hordel trades
Fixes #76

* Build legal move list
Randomize from that
Fixes #73

* Respect original number of moves when randomizing
Fixes #69

* Always pick at least one move
Fix bug with TM updating
Use new helpers to pick starters
Always update pokemon instance move set
Fix bug when selecting legal moves

* Fix Bonsly definition for moves and evolutions
Don't pick same starter as shadows
Respect shadow move flag when randomizing legal moves
Dont pick duplicate legal moves

* Update TM/Tutor Move compatibility based on legality if not changing compatibility and randomizing tutor moves
Fix Tutor Mover compatibility ordering
Respect ban shadow moves on TMs/Tutor Moves

* Don't randomize held item unless poke is holding an item
Fix some datatype size issues with pokespots
Use Xorshiro for tests

+++ Common/PokemonDefinitions/XD/XDMtBattlePokemon.cs
@@ -10,7 +10,7 @@ namespace XDCommon.PokemonDefinitions
     {
         const byte MtBattlePokemonSpeciesOffset = 0x02;
         const byte MtBattlePokemonMoveOffset = 0x06;
-        public byte Index { get; set; }
+        public int Index { get; set; }
 
         public ushort Exp { get; set; }
